[transport] Handle Thread Association request
Currently, NetworkProvisioning class, which takes part in
rendezvous over BLE session, only supports WiFi provisioning
and Thread provisioning is partially implemented in
RendezvousServer (Application layer). Move the Thread
provisioning code to the BLE transport layer to process
both network types in the same way.

Also reply to Thread Association request with SLAAC/On-Mesh
IPv6 Address once that is assigned so that the controller
knows that provisioning completed successfully and how to
reach the device.